requirements.txt
coloredlogs
dataclasses;python_version<"3.7"
librosa>=0.8.0
numpy>=1.18.3
pandas>=1.0.3
pocketsphinx
praat-textgrids==1.3.1
protobuf==3.19.4 # Protobuf 4.21.0 contains a breaking change for Python, as explained here: https://developers.google.com/protocol-buffers/docs/news/2022-05-06#python-updates
pyaudio
pydantic>=1.7.3
pytest==6.2.3
pytest-timeout==2.1.0
soundfile==0.10.3.post1
tensorboard==2.8.0
torch==1.10.1
torchaudio>=0.5.0
torchvision>=0.6.0
tqdm
